requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.3
gunicorn==22.0.0
pytest==8.3.2
pytest-cov==5.0.0
